requirements.txt
Flask==3.0.2
pandas==2.2.0
openpyxl==3.1.2
requests==2.31.0
beautifulsoup4==4.12.3
gunicorn==21.2.0
python-dotenv==1.0.1
pymongo==4.12.0
